Fixed some import issues
