+++ backend/ml_model/interfaces/data_point_creator_interface.py
@@ -31,10 +31,3 @@ class DataPointCreator(ABC):
             List: A list of DataPoint entities.
         """
         pass
-
-    @abstractmethod
-    def datapoint_creator(self) -> List:
-        """
-        Abstract method to create a DataPoint
-        """
-        pass
